project/requirements.txt
# HuBERT ASR Fine-tuning Dependencies

# Core ML
torch>=2.0.0
torchaudio>=2.0.0

# HuggingFace
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Metrics
jiwer>=3.0.0

# Configuration
pyyaml>=6.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0

# Logging & Visualization
tensorboard>=2.14.0
